+++ code/src/sixsq/nuvla/server/resources/module.clj
@@ -187,35 +187,31 @@ component, or application.
              (count creds))
       (throw (r/ex-response "Registries credentials can't be resolved!" 403)))))
 
+(defn throw-compatibility-required-for-application
+  [{:keys [subtype compatibility]}]
+  (when (and (utils/is-application? subtype) (nil? compatibility))
+    (throw (r/ex-response "Application subtype should have compatibility attribute set!" 400))))
+
 
 (defmethod crud/add resource-type
   [{:keys [body] :as request}]
 
   (a/throw-cannot-add collection-acl request)
-
   (throw-colliding-path (:path body))
-
   (throw-cannot-access-registries-or-creds request)
-
   (throw-price-error body)
+  (throw-compatibility-required-for-application body)
 
   (let [[{:keys [subtype] :as module-meta}
-         {:keys [author commit docker-compose] :as module-content}] (-> body u/strip-service-attrs
-                                                                        utils/split-resource)
-        module-meta (dissoc module-meta :compatibility :parent-path :published)]
+         {:keys [author commit] :as module-content}] (-> body u/strip-service-attrs
+                                                         utils/split-resource)
+        module-meta (dissoc module-meta :parent-path :published)]
 
     (if (utils/is-project? subtype)
       (db-add-module-meta module-meta request)
       (let [content-url     (subtype->resource-url subtype)
 
-            [compatibility
-             unsupported-options] (utils/parse-get-compatibility-fields subtype docker-compose)
-
-            content-body    (-> module-content
-                                (dissoc :unsupported-options)
-                                (merge {:resource-type content-url})
-                                (cond-> (seq unsupported-options) (assoc :unsupported-options
-                                                                         unsupported-options)))
+            content-body    (merge module-content {:resource-type content-url})
 
             content-request {:params      {:resource-name content-url}
                              :body        content-body
@@ -229,11 +225,7 @@ component, or application.
                                        :author author}
                                       commit (assoc :commit commit))])
             (set-price (auth/current-active-claim request))
-            (cond-> compatibility (assoc :compatibility compatibility))
             (db-add-module-meta request))))))
-
-
-
 
 
 (defn retrieve-edn
@@ -281,12 +273,12 @@ component, or application.
   (try
     (let [id          (str resource-type "/" (-> request :params :uuid))
           [module-meta
-           {:keys [author commit docker-compose] :as module-content}] (-> body
-                                                                          u/strip-service-attrs
-                                                                          utils/split-resource)
+           {:keys [author commit] :as module-content}] (-> body
+                                                           u/strip-service-attrs
+                                                           utils/split-resource)
           {:keys [subtype versions price acl]} (crud/retrieve-by-id-as-admin id)
           module-meta (-> module-meta
-                          (dissoc :compatibility :parent-path :published)
+                          (dissoc :parent-path :published)
                           (assoc :subtype subtype)
                           utils/set-parent-path)]
 
@@ -299,16 +291,7 @@ component, or application.
         (let [_              (throw-cannot-access-registries-or-creds request)
               content-url    (subtype->resource-url subtype)
 
-              [compatibility
-               unsupported-options] (when docker-compose
-                                      (utils/parse-get-compatibility-fields
-                                        subtype docker-compose))
-
-              content-body   (some-> module-content
-                                     (dissoc :unsupported-options)
-                                     (merge {:resource-type content-url})
-                                     (cond-> (seq unsupported-options) (assoc :unsupported-options
-                                                                              unsupported-options)))
+              content-body   (some-> module-content (merge {:resource-type content-url}))
 
               content-id     (when content-body
                                (-> {:params      {:resource-name content-url}
@@ -328,14 +311,14 @@ component, or application.
                                             (get-in module-meta [:price :cent-amount-daily]))
                                       (not= (:currency price)
                                             (get-in module-meta [:price :currency]))))]
-          (edit-impl
-            (assoc request
-              :body
-              (cond-> module-meta
-                      price-changed? (-> (assoc :price (merge price (:price module-meta)))
-                                         (set-price (auth/current-active-claim request)))
-                      versions (assoc :versions versions)
-                      compatibility (assoc :compatibility compatibility)))))))
+          (-> request
+              (update-in [:cimi-params :select] disj "compatibility")
+              (assoc :body
+                     (cond-> module-meta
+                             price-changed? (-> (assoc :price (merge price (:price module-meta)))
+                                                (set-price (auth/current-active-claim request)))
+                             versions (assoc :versions versions)))
+              edit-impl))))
     (catch Exception e
       (or (ex-data e) (throw e)))))